Moved to public repo to continue development.
Decided to move to a public repo as the goal of this program is to help others who prefer using 'docker run' as opposed to 'docker compose' for the deployment of their containers. I initially started this as a simple bash script to keep my docker containers up to date, but decided to continue working on it for a challenging project and to help others having the same issues I have.